Move live session error to public folder
