+++ FirebaseAI/Sources/Types/Internal/Live/LiveSessionService.swift
@@ -346,83 +346,3 @@ actor LiveSessionService {
     return AsyncWebSocket(urlSession: urlSession, urlRequest: urlRequest)
   }
 }
-
-/// The live model sent a message that the SDK failed to parse.
-///
-/// This may indicate that the SDK version needs updating, a model is too old for the current SDK
-/// version, or that the model is just
-/// not supported.
-///
-/// Check the `NSUnderlyingErrorKey` entry in ``errorUserInfo`` for the error that caused this.
-public struct LiveSessionUnsupportedMessageError: Error, Sendable, CustomNSError {
-  let underlyingError: Error
-
-  init(underlyingError: Error) {
-    self.underlyingError = underlyingError
-  }
-
-  public var errorUserInfo: [String: Any] {
-    [
-      NSLocalizedDescriptionKey: "Failed to parse a live message from the model. Cause: \(underlyingError.localizedDescription)",
-      NSUnderlyingErrorKey: underlyingError,
-    ]
-  }
-}
-
-/// The live session was closed, because the network connection was lost.
-///
-/// Check the `NSUnderlyingErrorKey` entry in ``errorUserInfo`` for the error that caused this.
-public struct LiveSessionLostConnectionError: Error, Sendable, CustomNSError {
-  let underlyingError: Error
-
-  init(underlyingError: Error) {
-    self.underlyingError = underlyingError
-  }
-
-  public var errorUserInfo: [String: Any] {
-    [
-      NSLocalizedDescriptionKey: "The live session lost connection to the server. Cause: \(underlyingError.localizedDescription)",
-      NSUnderlyingErrorKey: underlyingError,
-    ]
-  }
-}
-
-/// The live session was closed, but not for a reason the SDK expected.
-///
-/// Check the `NSUnderlyingErrorKey` entry in ``errorUserInfo`` for the error that caused this.
-public struct LiveSessionUnexpectedClosureError: Error, Sendable, CustomNSError {
-  let underlyingError: WebSocketClosedError
-
-  init(underlyingError: WebSocketClosedError) {
-    self.underlyingError = underlyingError
-  }
-
-  public var errorUserInfo: [String: Any] {
-    [
-      NSLocalizedDescriptionKey: "The live session was closed for some unexpected reason. Cause: \(underlyingError.localizedDescription)",
-      NSUnderlyingErrorKey: underlyingError,
-    ]
-  }
-}
-
-/// The live model refused our request to setup a live session.
-///
-/// This can occur due to the model not supporting the requested response modalities, the project
-/// not having access to the model,
-/// the model being invalid,  or some internal error.
-///
-/// Check the `NSUnderlyingErrorKey` entry in ``errorUserInfo`` for the error that caused this.
-public struct LiveSessionSetupError: Error, Sendable, CustomNSError {
-  let underlyingError: Error
-
-  init(underlyingError: Error) {
-    self.underlyingError = underlyingError
-  }
-
-  public var errorUserInfo: [String: Any] {
-    [
-      NSLocalizedDescriptionKey: "The model did not accept the live session request. Reason: \(underlyingError.localizedDescription)",
-      NSUnderlyingErrorKey: underlyingError,
-    ]
-  }
-}
